// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Rol = {
  USUARIO: 'USUARIO',
  ADMIN: 'ADMIN'
} as const

export type Rol = (typeof Rol)[keyof typeof Rol]


export const EstadoHabito = {
  ACTIVO: 'ACTIVO',
  PAUSADO: 'PAUSADO',
  ARCHIVADO: 'ARCHIVADO'
} as const

export type EstadoHabito = (typeof EstadoHabito)[keyof typeof EstadoHabito]


export const FrecuenciaHabito = {
  DIARIA: 'DIARIA',
  SEMANAL: 'SEMANAL',
  MENSUAL: 'MENSUAL'
} as const

export type FrecuenciaHabito = (typeof FrecuenciaHabito)[keyof typeof FrecuenciaHabito]